Split Scanner and bruteForceScanner in 2 file
